feat: Add isSetup call to expose if a user has a TOTP token. feat: Add optional per-user query to the isOptional and isEnabled settings (currently no effect)

+++ migration/1_add_auth_table.php
@@ -32,6 +32,9 @@ class AddAuthTable extends Horde_Db_Migration_Base
      */
     public function down()
     {
-        $this->dropTable(self::$tbl);
+        $tables = $this->tables();
+        if (in_array(self::$tbl, $tables)) {
+            $this->dropTable(self::$tbl);
+        }
     }
 }
